// src/modules/services/_apis/queries/protos/service-catalog/booking_catalog_service_query.ts
export interface GetAllBookingCatalogServicesRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue SerName = 1
     */
    serName?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.Int32Value ServPid = 2
     */
    servPid?: Int32Value;
    /**
     * @generated from protobuf field: optional google.protobuf.BoolValue IsOnlyChildren = 3
     */
    isOnlyChildren?: BoolValue;
    /**
     * @generated from protobuf field: optional PagingOption PagingOption = 4
     */
    pagingOption?: PagingOption;
    /**
     * @generated from protobuf field: repeated OrderByOption OrderBy = 5
     */
    orderBy: OrderByOption[];
}
/**
 * @generated from protobuf message BookingCatalogService.GetDetailBookingCatalogServiceRequest
 */
export interface GetDetailBookingCatalogServiceRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.Int32Value Id = 1
     */
    id?: Int32Value;
}
// ...
class GetAllBookingCatalogServicesRequest$Type extends MessageType<GetAllBookingCatalogServicesRequest> {
    constructor() {
        super("BookingCatalogService.GetAllBookingCatalogServicesRequest", [
            { no: 1, name: "SerName", kind: "message", jsonName: "SerName", T: () => StringValue },
            { no: 2, name: "ServPid", kind: "message", jsonName: "ServPid", T: () => Int32Value },
            { no: 3, name: "IsOnlyChildren", kind: "message", jsonName: "IsOnlyChildren", T: () => BoolValue },
            { no: 4, name: "PagingOption", kind: "message", jsonName: "PagingOption", T: () => PagingOption },
            { no: 5, name: "OrderBy", kind: "message", jsonName: "OrderBy", repeat: 2 /*RepeatType.UNPACKED*/, T: () => OrderByOption }
        ]);
    }
    create(value?: PartialMessage<GetAllBookingCatalogServicesRequest>): GetAllBookingCatalogServicesRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.orderBy = [];
        if (value !== undefined)
            reflectionMergePartial<GetAllBookingCatalogServicesRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetAllBookingCatalogServicesRequest): GetAllBookingCatalogServicesRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* optional google.protobuf.StringValue SerName */ 1:
                    message.serName = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.serName);
                    break;
                case /* optional google.protobuf.Int32Value ServPid */ 2:
                    message.servPid = Int32Value.internalBinaryRead(reader, reader.uint32(), options, message.servPid);
                    break;
                case /* optional google.protobuf.BoolValue IsOnlyChildren */ 3:
                    message.isOnlyChildren = BoolValue.internalBinaryRead(reader, reader.uint32(), options, message.isOnlyChildren);
                    break;
                case /* optional PagingOption PagingOption */ 4:
                    message.pagingOption = PagingOption.internalBinaryRead(reader, reader.uint32(), options, message.pagingOption);
                    break;
                case /* repeated OrderByOption OrderBy */ 5:
                    message.orderBy.push(OrderByOption.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetAllBookingCatalogServicesRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* optional google.protobuf.StringValue SerName = 1; */
        if (message.serName)
            StringValue.internalBinaryWrite(message.serName, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.Int32Value ServPid = 2; */
        if (message.servPid)
            Int32Value.internalBinaryWrite(message.servPid, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.BoolValue IsOnlyChildren = 3; */
        if (message.isOnlyChildren)
            BoolValue.internalBinaryWrite(message.isOnlyChildren, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* optional PagingOption PagingOption = 4; */
        if (message.pagingOption)
            PagingOption.internalBinaryWrite(message.pagingOption, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* repeated OrderByOption OrderBy = 5; */
        for (let i = 0; i < message.orderBy.length; i++)
            OrderByOption.internalBinaryWrite(message.orderBy[i], writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message BookingCatalogService.GetAllBookingCatalogServicesRequest
 */
export const GetAllBookingCatalogServicesRequest = new GetAllBookingCatalogServicesRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetDetailBookingCatalogServiceRequest$Type extends MessageType<GetDetailBookingCatalogServiceRequest> {
    constructor() {
        super("BookingCatalogService.GetDetailBookingCatalogServiceRequest", [
            { no: 1, name: "Id", kind: "message", jsonName: "Id", T: () => Int32Value }
        ]);
    }
    create(value?: PartialMessage<GetDetailBookingCatalogServiceRequest>): GetDetailBookingCatalogServiceRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<GetDetailBookingCatalogServiceRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetDetailBookingCatalogServiceRequest): GetDetailBookingCatalogServiceRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* optional google.protobuf.Int32Value Id */ 1:
                    message.id = Int32Value.internalBinaryRead(reader, reader.uint32(), options, message.id);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetDetailBookingCatalogServiceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* optional google.protobuf.Int32Value Id = 1; */
        if (message.id)
            Int32Value.internalBinaryWrite(message.id, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message BookingCatalogService.GetDetailBookingCatalogServiceRequest
 */
export const GetDetailBookingCatalogServiceRequest = new GetDetailBookingCatalogServiceRequest$Type();